refactor(db): split subscription repository into cohesive sub-modules
Behavior-preserving structural refactor of the 1240-line churn hotspot
backend/crates/db/src/repositories/subscription.rs. Converts the single
file into a subscription/ module whose mod.rs keeps the SubscriptionRepository
struct + constructor and module-level RLS docs, with the impl block split
across cohesive sub-modules (plans, subscriptions, payment_methods, invoices,
usage, events, coupons, statistics). Same proven pattern as the merged
document.rs split.

No public API changes: `pub mod subscription` + `pub use
subscription::SubscriptionRepository` re-export paths are unchanged, and every
method keeps its identical signature and body. Reduces the churn surface of any
single file.
